move middleware to utils
