+++ src/litserve/utils.py
@@ -18,6 +18,8 @@ import uuid
 from typing import Coroutine, Optional
 
 from fastapi import HTTPException
+from starlette.middleware.base import BaseHTTPMiddleware
+from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
 logger = logging.getLogger(__name__)
 
@@ -65,3 +67,32 @@ async def azip(*async_iterables):
         if any(isinstance(result, StopAsyncIteration) for result in results):
             break
         yield tuple(results)
+
+
+class MaxSizeMiddleware(BaseHTTPMiddleware):
+    def __init__(
+        self,
+        app: ASGIApp,
+        *,
+        max_size: Optional[int] = None,
+    ) -> None:
+        self.app = app
+        self.max_size = max_size
+
+    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
+        if scope["type"] != "http":
+            await self.app(scope, receive, send)
+            return
+
+        total_size = 0
+
+        async def rcv() -> Message:
+            nonlocal total_size
+            message = await receive()
+            chunk_size = len(message.get("body", b""))
+            total_size += chunk_size
+            if self.max_size is not None and total_size > self.max_size:
+                raise HTTPException(413, "Payload too large")
+            return message
+
+        await self.app(scope, rcv, send)
